requirements.txt
openpyxl
pyyaml
requests
mysql-connector-python
argparse
setuptools
wheel
tqdm